fix(server/pwa): move to middleware
