+++ scripts/build-service-workers.js
@@ -23,7 +23,7 @@ const babel = require('rollup-plugin-babel');
 
 async function buildServiceWorkerScripts({ dev = false, minify = true } = {}) {
   const inputDirectory = path.resolve(__dirname, '../src/client/service-worker');
-  const buildFolderDirectory = path.resolve(__dirname, '../lib/server/pwa/middleware', 'scripts');
+  const buildFolderDirectory = path.resolve(__dirname, '../lib/server/middleware/pwa', 'scripts');
 
   const plugins = [
     resolve(),
