--- src/server/pwa/createRouter.js
@@ -1,34 +1,0 @@
-/*
- * Copyright 2020 American Express Travel Related Services Company, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-import express from 'express';
-
-import { serviceWorkerMiddleware } from './middleware/service-worker';
-
-// used to reference the paths used by the router
-export const routes = {
-  prefix: '/_/pwa',
-  worker: '/service-worker.js',
-};
-
-export function createPWARouter(routeNames = routes) {
-  const pwaRouter = express.Router();
-
-  pwaRouter
-    .get(routeNames.worker, serviceWorkerMiddleware());
-
-  return pwaRouter;
-}
