+++ src/server/ssrServer.js
@@ -45,7 +45,7 @@ import sendHtml, { renderStaticErrorPage } from './middleware/sendHtml';
 import logging from './utils/logging/serverMiddleware';
 import forwardedHeaderParser from './middleware/forwardedHeaderParser';
 
-import { createPWARouter } from './pwa';
+import { serviceWorkerMiddleware } from './middleware/pwa';
 
 export function createApp({ enablePostToModuleRoutes = false } = {}) {
   const app = express();
@@ -58,8 +58,8 @@ export function createApp({ enablePostToModuleRoutes = false } = {}) {
   app.use(forwardedHeaderParser);
 
   app.use('/_/static', express.static(path.join(__dirname, '../../build'), { maxage: '182d' }));
+  app.get('/_/pwa/service-worker.js', serviceWorkerMiddleware());
   app.get('*', addCacheHeaders);
-  app.use('/_/pwa', createPWARouter());
 
   app.disable('x-powered-by');
   app.disable('e-tag');
